requirements.txt
# Core Dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0
streamlit>=1.22.0

# Configuration
pyyaml>=6.0

# Network Analysis (Optional - for PCAP processing)
scapy>=2.5.0

# Model Export (Optional)
onnx>=1.14.0
onnxruntime>=1.15.0

# Development Tools
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
jupyter>=1.0.0

# Documentation
sphinx>=6.2.0
sphinx-rtd-theme>=1.2.0

# Performance
psutil>=5.9.0
